Gather function to work with result in dedicated trait
